factor out shared code between cli tools
