+++ java/src/main/java/com/mlt/cli/Decode.java
@@ -1,9 +1,7 @@
 package com.mlt.tools;
 
-import com.mlt.data.MapLibreTile;
 import com.mlt.decoder.MltDecoder;
 import com.mlt.metadata.tileset.MltTilesetMetadata;
-import com.mlt.vector.FeatureTable;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import org.apache.commons.cli.CommandLine;
@@ -14,31 +12,6 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
 public class Decode {
-
-  public static void printMLT(MapLibreTile mlTile) {
-    var mltLayers = mlTile.layers();
-    for (var i = 0; i < mltLayers.size(); i++) {
-      var mltLayer = mltLayers.get(i);
-      System.out.println(mltLayer.name());
-      var mltFeatures = mltLayer.features();
-      for (var j = 0; j < mltFeatures.size(); j++) {
-        var mltFeature = mltFeatures.get(j);
-        System.out.println("  " + mltFeature);
-      }
-    }
-  }
-
-  private static void printMLTVectorized(FeatureTable[] featureTables) {
-    for (var i = 0; i < featureTables.length; i++) {
-      var featureTable = featureTables[i];
-      System.out.println(featureTable.getName());
-      var featureIterator = featureTable.iterator();
-      while (featureIterator.hasNext()) {
-        var mltFeature = featureIterator.next();
-        System.out.println("  " + mltFeature);
-      }
-    }
-  }
 
   private static final String FILE_NAME_ARG = "mlt";
   private static final String PRINT_MLT_OPTION = "printmlt";
@@ -98,16 +71,21 @@ public class Decode {
 
       Timer timer = new Timer();
       if (willUseVectorized) {
-        var decodedTile = MltDecoder.decodeMlTileVectorized(mltTileBuffer, tileMetadata);
+        var featureTables = MltDecoder.decodeMlTileVectorized(mltTileBuffer, tileMetadata);
+        // Note: the vectorized result is a FeatureTable array
+        // which provides an iterator to access the features.
+        // Therefore, we must iterate over the FeatureTable array
+        // to trigger actual decoding of the features.
+        CliUtil.decodeFeatureTables(featureTables);
         if (willTime) timer.stop("decoding");
         if (willPrintMLT) {
-          printMLTVectorized(decodedTile);
+          CliUtil.printMLTVectorized(featureTables);
         }
       } else {
         var decodedTile = MltDecoder.decodeMlTile(mltTileBuffer, tileMetadata);
         if (willTime) timer.stop("decoding");
         if (willPrintMLT) {
-          printMLT(decodedTile);
+          CliUtil.printMLT(decodedTile);
         }
       }
     } catch (Exception e) {
